default to NFTs on Collection Page, misc minor improvements
