+++ compositions/Collections/CollectionNav.css.ts
@@ -7,3 +7,9 @@ export const activeLine = style([
     bottom: -2,
   },
 ])
+
+export const navItem = style([
+  {
+    userSelect: 'none',
+  },
+])
